refactor(contracts): extract FakeWorkers to roxabi-nats[testing] (ADR-059 V6)
Resolves the circular dep contracts[testing] → roxabi-nats → contracts by
moving FakeTtsWorker/FakeSttWorker/FakeImageWorker to their natural owner,
the NATS transport package.

- Add roxabi_nats/testing/ package: voice.py, image.py, _guards.py, __init__.py
- Replace contracts voice/image testing.py with backward-compat re-export shims
- Migrate _testing_guards.py to roxabi_nats/testing/_guards.py (re-export shim in contracts)
- Move test_voice_testing_doubles.py + test_image_testing_doubles.py to roxabi-nats/tests/
- roxabi-nats[testing] = ["roxabi-contracts[testing]"] (new optional-deps)
- roxabi-contracts[testing] removes nats-py + roxabi-nats, keeps scipy only

Closes #962
