+++ packages/roxabi-contracts/pyproject.toml
@@ -9,11 +9,9 @@ authors = [{ name = "Roxabi SAS", email = "mickael@bouly.io" }]
 dependencies = ["pydantic>=2"]
 
 [project.optional-dependencies]
-# Workspace-only: roxabi-nats is not published to PyPI (ADR-045 §PyPI deferral).
-# The bare name resolves inside the uv workspace via lockfile editable reference.
-# External consumers installing roxabi-contracts[testing] from a git tag must
-# declare their own [tool.uv.sources] roxabi-nats = { git = ... } entry.
-testing = ["nats-py>=2.6", "roxabi-nats", "scipy>=1.11"]
+# scipy + numpy are required by roxabi_contracts.voice.fixtures (silence_wav_16khz).
+# nats-py and roxabi-nats have moved to roxabi-nats[testing] (ADR-059 V6).
+testing = ["scipy>=1.11"]
 
 [project.urls]
 Repository = "https://github.com/Roxabi/lyra"
