+++ packages/roxabi-nats/pyproject.toml
@@ -19,6 +19,12 @@ dependencies = [
     "roxabi-contracts>=0.1.0",
 ]
 
+[project.optional-dependencies]
+# scipy + numpy are pulled in via roxabi-contracts[testing] (voice fixtures).
+# nats-py is already a core dep — listed here for install-time discoverability.
+# ADR-059 V6: FakeTtsWorker/FakeSttWorker/FakeImageWorker live in roxabi_nats.testing.
+testing = ["roxabi-contracts[testing]"]
+
 [project.urls]
 Repository = "https://github.com/Roxabi/lyra"
 Issues = "https://github.com/Roxabi/lyra/issues"
